screenio refactoring + c89 fix for [r5045]
libcob:
* screenio.c:
  * (cob_screen_attr): always parse COBOL and curses color value
  * (cob_to_curses_color): signature changed 
     * operate on integer to be reusable for CONTROL handling
     * separate return value instead of setting the default color
  * (adjust_attr_from_extended_color): new function directly handling the attribute adjustment, replaced (has_extended_color)
  * (handle_control_field_color, get_cob_color_from_color_value): new functions to get COBOL color number from text and handling it
* common.h (cob_colors): changed color defines to enum
* move.c: fix some compiler warnings by explicit casts

+++ libcob/move.c
@@ -486,7 +486,7 @@ cob_move_display_to_packed (cob_field *f1, cob_field *f2)
 		   the half-byte before setting the sign */
 
 		/* check for necessary loop (until we not need the p_end check) */
-		if (i_end - i < (p_end - p + 1) / 2) {
+		if (i_end - i < (unsigned int)(p_end - p + 1) / 2) {
 			while (i <= i_end) {
 				*q = (unsigned char) (*p << 4)	/* -> dropping the higher bits = no use in COB_D2I */
 					+ COB_D2I (*(p + 1));
@@ -1548,7 +1548,7 @@ cob_move (cob_field *src, cob_field *dst)
 							val /= cob_exp10_ll[diff_scale];
 						}
 						if (val >= INT_MIN && val <= INT_MAX) {
-							cob_set_packed_int (dst, val);
+							cob_set_packed_int (dst, (int)val);
 							return;
 						}
 					}
